Move search reducers and context
