+++ src/pages/search/dashboard/Ecosystems.tsx
@@ -12,7 +12,7 @@ import EcosystemsBox from "pages/search/dashboard/ecosystems/EcosystemsBox";
 import {
   SearchLegacyCTX,
   type LegacyContextValues,
-} from "pages/search/SearchContext";
+} from "pages/search/hooks/SearchContext";
 import formatNumber from "pages/search/utils/format";
 import matchColor from "pages/search/utils/matchColor";
 import BackendAPI from "pages/search/api/backendAPI";
