+++ src/pages/search/dashboard/Species.jsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import Accordion from "pages/search/layout/Accordion";
 import Richness from "pages/search/dashboard/species/Richness";
-import { SearchLegacyCTX } from "../SearchContext";
+import { SearchLegacyCTX } from "../hooks/SearchContext";
 
 class Species extends React.Component {
   static contextType = SearchLegacyCTX;
