+++ src/pages/search/dashboard/landscape/CompensationFactor.tsx
@@ -10,7 +10,7 @@ import TextBoxes from "pages/search/shared_components/TextBoxes";
 import {
   SearchLegacyCTX,
   type LegacyContextValues,
-} from "pages/search/SearchContext";
+} from "pages/search/hooks/SearchContext";
 import { cfData } from "pages/search/types/compensationFactor";
 import { textsObject } from "pages/search/types/texts";
 import LargeStackedBar from "pages/search/shared_components/charts/LargeStackedBar";
