+++ src/pages/search/dashboard/landscape/connectivity/CurrentSEPAConnectivity.tsx
@@ -7,7 +7,7 @@ import { IconTooltip } from "pages/search/shared_components/Tooltips";
 import {
   SearchLegacyCTX,
   type LegacyContextValues,
-} from "pages/search/SearchContext";
+} from "pages/search/hooks/SearchContext";
 import formatNumber from "pages/search/utils/format";
 import matchColor from "pages/search/utils/matchColor";
 import BackendAPI from "pages/search/api/backendAPI";
