+++ src/pages/search/dashboard/landscape/humanFootprint/CurrentFootprint.tsx
@@ -5,7 +5,7 @@ import InfoIcon from "@mui/icons-material/Info";
 import {
   SearchLegacyCTX,
   type LegacyContextValues,
-} from "pages/search/SearchContext";
+} from "pages/search/hooks/SearchContext";
 import { ShortInfo } from "core/composites/ShortInfo";
 import { IconTooltip } from "pages/search/shared_components/Tooltips";
 import matchColor from "pages/search/utils/matchColor";
