+++ src/pages/search/layout/Dashboard.tsx
@@ -5,7 +5,7 @@ import Paisaje from "@mui/icons-material/FilterHdr";
 // NOTE: No consumido en esta implementación
 // import Especies from "@mui/icons-material/FilterVintage";
 
-import { useSearchStateCTX } from "pages/search/SearchContext";
+import { useSearchStateCTX } from "pages/search/hooks/SearchContext";
 import Landscape from "pages/search/dashboard/Landscape";
 import Species from "pages/search/dashboard/Species";
 import Ecosystems from "pages/search/dashboard/Ecosystems";
