+++ src/pages/search/layout/MapViewer.tsx
@@ -15,7 +15,7 @@ import CloseIcon from "@mui/icons-material/Close";
 
 import { DrawControl } from "pages/search/mapViewer/DrawControl";
 import type { Polygon as PolygonType } from "pages/search/types/dashboard";
-import { useSearchStateCTX } from "pages/search/SearchContext";
+import { useSearchStateCTX } from "pages/search/hooks/SearchContext";
 import "leaflet/dist/leaflet.css";
 import { useUserCTX } from "core/hooks/UserContext";
 import { COLOMBIA_BOUNDS } from "pages/utils/settings";
