+++ src/pages/search/selector/DrawPolygon.tsx
@@ -10,7 +10,7 @@ import RemoveIcon from "pages/search/selector/RemoveIcon";
 import {
   useSearchDrawControlsCTX,
   useSearchLegacyCTX,
-} from "pages/search/SearchContext";
+} from "pages/search/hooks/SearchContext";
 import { uiText } from "pages/search/selector/drawPolygon/uiText";
 import "pages/search/layout/DrawPolygon.css";
 import { DrawMode } from "pages/search/selector/drawPolygon/types";
